+++ src/components/Inputs/AutocompleteFilter/index.ts
@@ -1,0 +1,4 @@
+import Autocomplete from "./AutocompleteFilter";
+
+export default Autocomplete;
+export * from "./AutocompleteFilter";