+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.4.2
-- **[fix]** - `Chip` forward ref
+## v2.5.0
+- **[feat]** - Add new component `AutocompleteFilter`
+- **[chore]** - upgrade storybook to v8.3.6